add more test cases improve harmony-duplicate-export test cases to allow scope-hoisting
