+++ test/cases/parsing/harmony-duplicate-export/cjs/1.js
