+++ test/cases/parsing/harmony-duplicate-export/cjs/2.js
