+++ test/cases/parsing/harmony-duplicate-export/cjs/3.js
