+++ test/cases/parsing/harmony-duplicate-export/cjs/4.js
