+++ test/cases/parsing/harmony-duplicate-export/cjs/5.js
