+++ test/cases/parsing/harmony-duplicate-export/cjs/6.js
