+++ test/cases/parsing/harmony-duplicate-export/cjs/7.js
