+++ test/cases/parsing/harmony-duplicate-export/cjs/a.js
@@ -1,0 +1,1 @@
+export var x = "a";
