+++ test/cases/parsing/harmony-duplicate-export/cjs/b.js
@@ -1,0 +1,1 @@
+export var x = "b";
