+++ test/cases/parsing/harmony-duplicate-export/cjs/c.js
@@ -1,0 +1,1 @@
+exports.x = "c";
