+++ test/cases/parsing/harmony-duplicate-export/cjs/d.js
@@ -1,0 +1,1 @@
+exports.x = "d";
