+++ test/cases/scope-hoisting/reexport-star-exposed-cjs/a.js
@@ -1,0 +1,1 @@
+exports.named = "named";
