+++ test/cases/scope-hoisting/reexport-star-exposed-cjs/b.js
@@ -1,0 +1,1 @@
+export { named } from "./a";
