+++ test/cases/scope-hoisting/reexport-star-exposed-cjs/c.js
@@ -1,0 +1,1 @@
+export * from "./b";
