+++ test/cases/scope-hoisting/reexport-star-exposed-cjs/index.js
@@ -1,0 +1,5 @@
+var c = require("./c");
+
+it("should have the correct values", function() {
+	expect(c.named).toBe("named");
+});
