+++ test/cases/scope-hoisting/reexport-star-external-cjs/a.js
@@ -1,0 +1,1 @@
+exports.named = "named";
