+++ test/cases/scope-hoisting/reexport-star-external-cjs/b.js
@@ -1,0 +1,1 @@
+export var other = "other";
