+++ test/cases/scope-hoisting/reexport-star-external-cjs/c.js
@@ -1,0 +1,2 @@
+export * from "./a";
+export * from "./b";
